+++ tests/end2end/features/image/test_write_bdd.py
@@ -35,7 +35,10 @@ def exif_content():
 @bdd.when("I write the image to <name>")
 def write_image(handler, name):
     handler.write_pixmap(
-        handler.current, path=name, original_path=handler._path, parallel=False
+        handler._edit_handler.pixmap,
+        path=name,
+        original_path=handler._path,
+        parallel=False,
     )