+++ tests/unit/imutils/test_imtransform.py
@@ -7,40 +7,12 @@
 """Tests for vimiv.imutils.imtransform."""
 
 from functools import partial
-from unittest import mock
 
-from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QPixmap
 
 import pytest
 
-from vimiv.imutils import imtransform
-
-
-class MockHandler(mock.MagicMock):
-    """Stub used as file handler for Transform.
-
-    Provides a generic pixmap for all pixmaps and is always editable.
-    """
-
-    def __init__(self):
-        super().__init__()
-        self._pixmap = QPixmap(300, 300)
-        self._pixmap.fill(Qt.black)
-
-    @property
-    def editable(self):
-        return True
-
-    @property
-    def transformed(self):
-        return self._pixmap
-
-    @transformed.setter
-    def transformed(self, value):
-        pass
-
-    original = current = transformed
+from vimiv.imutils import current_pixmap, imtransform
 
 
 ACTIONS = (
@@ -52,10 +24,13 @@ ACTIONS = (
 
 
 @pytest.fixture(scope="function")
-def transform(qtbot):
+def transform(qtbot, mocker):
     """Fixture to retrieve a clean Transform instance."""
-    transform.handler = MockHandler()  # Keep as reference here due to weakref
-    return imtransform.Transform(transform.handler)
+    pixmap = QPixmap(300, 300)
+    current_pm = current_pixmap.CurrentPixmap()
+    transform = imtransform.Transform(current_pm)
+    current_pm.pixmap = transform.original = pixmap
+    return transform
 
 
 @pytest.fixture(params=ACTIONS)