+++ vimiv/imutils/imtransform.py
@@ -8,11 +8,10 @@
 
 import functools
 import math
-import weakref
 from typing import Optional
 
-from PyQt5.QtCore import Qt, QRect, QSize
-from PyQt5.QtGui import QTransform
+from PyQt5.QtCore import Qt, QRect, QSize, QObject, pyqtSignal
+from PyQt5.QtGui import QTransform, QPixmap
 
 from vimiv import api
 from vimiv.utils import log
@@ -45,13 +44,39 @@ class Transform(QTransform):
     to apply these transformations to the pixmap given by the handler.
 
     Attributes:
-        _handler: weak reference to ImageFileHandler used to retrieve/set updated files.
+        _current: Class to access the currently displayed pixmap.
+        _original: The original, untransformed, pixmap.
+
+    Signals:
+        transformed: Emitted with the transformed pixmap upon changes.
     """
 
+    class Signals(QObject):
+        """Signals for transformed required as QTransform is not a QObject."""
+
+        transformed = pyqtSignal(QPixmap)
+
+    _signals = Signals()
+    transformed = _signals.transformed
+
     @api.objreg.register
-    def __init__(self, handler):
+    def __init__(self, current_pixmap):
         super().__init__()
-        self._handler = weakref.ref(handler)
+        self._current = current_pixmap
+        self._original = None
+
+    @property
+    def current(self):
+        return self._current.pixmap
+
+    @property
+    def original(self):
+        return self._original
+
+    @original.setter
+    def original(self, pixmap):
+        self._original = pixmap
+        self.reset()
 
     @property
     def angle(self) -> float:
@@ -107,8 +132,8 @@ class Transform(QTransform):
 
         .. note:: This transforms the original image and writes to disk.
         """
-        dx = width / self._handler().transformed.width()
-        dy = dx if height is None else height / self._handler().transformed.height()
+        dx = width / self.current.width()
+        dy = dx if height is None else height / self.current.height()
         self.scale(dx, dy)
 
     @register_transform_command()
@@ -129,8 +154,7 @@ class Transform(QTransform):
 
     def apply(self):
         """Apply all transformations to the original pixmap."""
-        original = self._handler().original
-        self._apply(original.transformed(self, mode=Qt.SmoothTransformation))
+        self._apply(self.original.transformed(self, mode=Qt.SmoothTransformation))
 
     def straighten(self, *, angle: int, original_size: QSize):
         """Straighten the original image.
@@ -142,9 +166,8 @@ class Transform(QTransform):
             angle: Rotation angle to straighten the original image by.
             original_size: Size of the original unstraightened image.
         """
-        original = self._handler().original
         self.rotate(angle)
-        transformed = original.transformed(self, mode=Qt.SmoothTransformation)
+        transformed = self.original.transformed(self, mode=Qt.SmoothTransformation)
         rect = self.largest_rect_in_rotated(
             original=original_size, rotated=transformed.size(), angle=angle
         )
@@ -157,10 +180,10 @@ class Transform(QTransform):
                 "Error transforming image, ignoring transformation.\n"
                 "Is the resulting image too large? Zero?."
             )
-        self._handler().transformed = transformed
+        self.transformed.emit(transformed)
 
     def _ensure_editable(self):
-        if not self._handler().editable:
+        if not self._current.editable:
             raise api.commands.CommandError("File format does not support transform")
 
     @property
@@ -182,13 +205,13 @@ class Transform(QTransform):
     @property
     def size(self) -> QSize:
         """Size of the transformed image."""
-        return self._handler().transformed.size()
+        return self.current.size()
 
     @api.commands.register(mode=api.modes.IMAGE)
     def undo_transformations(self):
         """Undo any transformation applied to the current image."""
         self.reset()
-        self._handler().transformed = self._handler().original
+        self.transformed.emit(self.original)
 
     @classmethod
     def largest_rect_in_rotated(